feat: move Network class to new network module separate from player
