+++ energetica/database/player.py
@@ -19,7 +19,6 @@ from energetica.database import db
 from energetica.database.active_facility import ActiveFacility
 from energetica.database.engine_data import (
     CapacityData,
-    CircularBufferNetwork,
     CircularBufferPlayer,
     CumulativeEmissionsData,
 )
@@ -752,27 +751,6 @@ class Player(db.Model, UserMixin):
             if technologies:
                 pages_data |= {"technologies": self.cache.technologies_data}
             self.emit("update_page_data", pages_data)
-
-
-@dataclass
-class NetworkData:
-    """Dataclass that stores the network data."""
-
-    rolling_history: CircularBufferNetwork = field(default_factory=CircularBufferNetwork)
-    capacities: CapacityData = field(default_factory=CapacityData)
-
-
-class Network(db.Model):
-    """Class that stores the networks of players."""
-
-    id = db.Column(db.Integer, primary_key=True)
-    name = db.Column(db.String(50), unique=True)
-    members = db.relationship("Player", backref="network")
-
-    @cached_property
-    def data(self) -> NetworkData:
-        """Cached property that stores the network data."""
-        return current_app.config["engine"].data["by_network"][self.id]
 
 
 class PlayerUnreadMessages(db.Model):
